Move user validation to registration step to catch errors before send…

+++ server/src/controllers/userRegister.js
@@ -1,6 +1,6 @@
 import sendEmail from "../util/mailer.js";
 import generateCode from "../util/codeGenerator.js";
-import User from "../models/User.js";
+import User, { validateUser } from "../models/User.js";
 import PendingUser from "../models/PendingUser.js";
 import bcrypt from "bcrypt";
 import { logError } from "../util/logging.js";
@@ -25,6 +25,22 @@ export const userRegister = async (req, res) => {
     }
 
     const hashedPassword = await bcrypt.hash(password, SALT_ROUNDS);
+
+    const userData = {
+      email,
+      username,
+      profile: {
+        first_name: firstName,
+        last_name: lastName,
+      },
+      password: hashedPassword,
+    };
+
+    const validationErrors = validateUser(userData);
+    if (validationErrors.length > 0) {
+      return res.status(400).json({ errors: validationErrors });
+    }
+
     const verificationCode = generateCode();
 
     await PendingUser.deleteOne({ email });
